mcp/practice/langchain-mcp-adapters/requirements.txt
langchain-mcp-adapters==0.2.1
langgraph==1.0.4
langchain==1.1.3
mcp==1.24.0
python-dotenv